requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
django-jazzmin==2.6.1
djangorestframework==3.15.1
django-tinymce
django-tinymce4-lite
fuzzywuzzy==0.18.0
idna==3.6
mysqlclient==2.2.4
numpy==1.26.4
pandas==2.2.1
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
